+++ packages/create-node-app-core/loaders.ts
@@ -80,6 +80,8 @@ type FileLoaderOptions = {
   appName: string;
   originalDirectory: string;
   verbose: boolean;
+  useYarn?: boolean;
+  usePnp?: boolean;
   srcDir: string;
   mode?: string;
   runCommand: string;
@@ -97,7 +99,9 @@ const copyLoader: FileLoader =
   ({ path }) => {
     return copyFile(
       `${templateDir}/${path}`,
-      `${root}/${path}`.replace(SRC_PATH_PATTERN, getSrcDirPattern(srcDir)),
+      `${root}/${path}`
+        .replace(/.if-(npm|yarn|pnpm)$/, "")
+        .replace(SRC_PATH_PATTERN, getSrcDirPattern(srcDir)),
       verbose
     );
   };
@@ -107,6 +111,7 @@ const appendLoader: FileLoader =
   ({ path }) => {
     const newPath = path
       .replace(/.append$/, "")
+      .replace(/.if-(npm|yarn|pnpm)$/, "")
       .replace(SRC_PATH_PATTERN, getSrcDirPattern(srcDir));
     return appendFile(`${templateDir}/${path}`, `${root}/${newPath}`, verbose);
   };
@@ -130,6 +135,7 @@ const templateLoader: FileLoader =
     const newPath = path
       .replace(/.template$/, "")
       .replace(/.append$/, "")
+      .replace(/.if-(npm|yarn|pnpm)$/, "")
       .replace(SRC_PATH_PATTERN, getSrcDirPattern(srcDir));
 
     return writeFile(
@@ -153,6 +159,8 @@ const fileLoader: FileLoader =
     appName,
     originalDirectory,
     verbose,
+    useYarn,
+    usePnpm,
     srcDir = DEFAULT_SRC_PATH,
     runCommand,
     installCommand,
@@ -174,6 +182,8 @@ const fileLoader: FileLoader =
       appName,
       originalDirectory,
       verbose,
+      useYarn,
+      usePnpm,
       mode,
       srcDir,
       runCommand,
@@ -192,6 +202,8 @@ export type LoadFilesOptions = {
   appName: string;
   originalDirectory: string;
   verbose: boolean;
+  useYarn?: boolean;
+  usePnpm?: boolean;
   srcDir?: string;
   runCommand: string;
   installCommand: string;
@@ -205,6 +217,8 @@ export const loadFiles = async ({
   appName,
   originalDirectory,
   verbose,
+  useYarn = false,
+  usePnpm = false,
   srcDir = DEFAULT_SRC_PATH,
   runCommand,
   installCommand,
@@ -225,6 +239,15 @@ export const loadFiles = async ({
         "!template.json",
         "!yarn.lock",
         "!pnpm-lock.yaml",
+        // based on the package manager we want to ignore files containing
+        // the other package as condition.
+        // For example, if `usePnpm` is true, the we need to ignore
+        // all files with `.if-npm` or `.if-yarn` somewhere in the name.
+        ...(usePnpm
+          ? ["!*.if-npm.*", "!*.if-yarn.*"]
+          : useYarn
+          ? ["!*.if-npm.*", "!*.if-pnpm.*"]
+          : ["!*.if-yarn.*", "!*.if-pnpm.*"]),
       ],
       directoryFilter: ["!package"],
     })) {
@@ -235,6 +258,8 @@ export const loadFiles = async ({
           appName,
           originalDirectory,
           verbose,
+          useYarn,
+          usePnpm,
           srcDir,
           runCommand,
           installCommand,