+++ packages/create-node-app-core/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @create-node-app/core
 
+## 0.3.9
+
+### Patch Changes
+
+- Small fix
+
+## 0.3.8
+
+### Patch Changes
+
+- Simple fix
+
+## 0.3.7
+
+### Patch Changes
+
+- Updated package loader
+
+## 0.3.6
+
+### Patch Changes
+
+- Added conditional installation for npm, yarn and pnpm
+
 ## 0.3.5
 
 ### Patch Changes